Implement current temperature feature

+++ temperature/models.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, ForeignKey, DateTime
+from sqlalchemy import Column, Integer, ForeignKey, DateTime, Float
 from sqlalchemy.orm import relationship
 from city import models
 
@@ -11,5 +11,5 @@ class DBTemperature(Base):
     id = Column(Integer, primary_key=True, index=True)
     city_id = Column(Integer, ForeignKey("cities.id"))
     date_time = Column(DateTime)
-    temperature = Column(Integer)
+    temperature = Column(Float)
     city = relationship("DBCity", back_populates="temperature")
